+++ app/src/main/java/com/me/recipe/ui/navigation/BottomNavigationScreens.kt
@@ -1,0 +1,3 @@
+package com.me.recipe.ui.navigation
+
+val bottomNavigationScreens = listOf(HomeDestination, RecipeListDestination)
